hackernews-viewer/frontend/pyproject.toml
[tool.poetry]
name = "hackernews-viewer-frontend"
version = "0.1.0"
description = "Frontend for HackerNews Viewer application"
authors = ["Devin AI <devin-ai-integration[bot]@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
streamlit = "^1.31.0"
httpx = "^0.26.0"
python-dotenv = "^1.0.0"
pandas = "^2.1.0"
plotly = "^5.18.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^23.7.0"
isort = "^5.12.0"
mypy = "^1.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
